Initial Path Planning & Cleanliness

+++ plc_sketchbook/stage_control/a_declarations.ino
@@ -17,15 +17,15 @@ void clearSerialBuffer();
 
 // Stage Motion
 void set_stage_power(bool state);
-void stage(int requested_position);
+void stage(float requested_position);
 
 // Distance Sensor
 void set_distance_sensor_power(bool state);
-float read_distance();
+uint32_t read_distance();
 
 // Data Maintenance
 void update_path_history(float current_position);
-float calculate_path();
+uint32_t calculate_path();
 
 // CPU I/O
 String status_check(uint8_t Slot);
